requirements.txt
pandas
openai>=1.0
python-dotenv

